requirements.txt
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.45.5
certifi @ file:///croot/certifi_1745939216646/work/certifi
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
fastapi==0.115.12
fastchat==0.1.0
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.0
git-lfs==1.6
gitdb==4.0.12
GitPython==3.1.44
gradio==4.16.0
gradio_client==0.8.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
latex2mathml==3.77.0
lazy_loader==0.4
-e git+https://github.com/haotian-liu/LLaVA.git@c121f0432da27facab705978f83c4ada465e46fd#egg=llava
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
narwhals==1.35.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
openai==1.78.0
orjson==3.10.16
packaging==25.0
pandas==2.2.3
peft==0.15.2
pillow==10.4.0
platformdirs==4.3.8
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.2.2
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.28.0
setproctitle==1.3.6
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
svgwrite==1.4.3
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==0.6.13
tokenizers==0.13.3
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.31.0
triton==2.1.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wandb==0.19.11
wavedrom==2.0.3.post3
websockets==11.0.3
xxhash==3.5.0
yarl==1.20.0
